+++ src/plugin/fixes/disable-per-file.test.ts
@@ -1,4 +1,4 @@
-import { TransformTester } from '../../test-util/transform-tester.js';
+import { TransformTester } from '../../test-util/fix-tester.js';
 
 const tester = new TransformTester(
   'disablePerFile',
