+++ src/plugin/fixes/disable-per-line.test.ts
@@ -1,4 +1,4 @@
-import { TransformTester } from '../../test-util/transform-tester.js';
+import { TransformTester } from '../../test-util/fix-tester.js';
 
 const tester = new TransformTester(
   'disablePerLine',
